+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed white space issue in parsing forces block in .castep file.
 - Fixed bug in parsing .castep file when geometry <frequency> is unchanged from its initial value.
 - Fixed bug in parsing .castep file when there is an initial SCF cycle on continuation of a geometry optimisation.
+- Only add final info to run `dict` if the final info includes the total time string.
 
 ## [0.2.0] - 2019.11.22