+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed bug in parsing .castep file when there is an initial SCF cycle on continuation of a geometry optimisation.
 - Only add final info to run `dict` if the final info includes the total time string.
 - Fixed issue when adding up total time if no previous SCF cycle to get last time from.
+- Skip runs if they don't include the string "Cell Contents"
 
 ## [0.2.0] - 2019.11.22